moving IdentityOverview to display/fewlines folder
